chore(dashboard-prototype): move the testing menu to the top of the Dashboard page
